backend/requirements.txt
fastapi>=0.116.1
uvicorn>=0.35.0
python-dotenv>=1.0.1
langchain>=0.3.27
langchain-core>=0.3.74
langchain-groq>=0.3.7
langgraph>=0.6.6
